pyproject.toml
[project]
name = "chat-mcp"
version = "0.1.0"
description = "Multi-agent collaborative chat system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "sse-starlette>=2.0.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    "textual>=0.80.0",
    "pyyaml>=6.0.0",
    "python-multipart>=0.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
chat = "chat_mcp.cli.main:cli"
chat-server = "chat_mcp.server.app:main"
chat-agent = "chat_mcp.agent.runner:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
